+++ pyproject.toml
@@ -3,7 +3,7 @@ name = "ai-traineree"
 version = "0.9.0"
 description = "Yet another zoo of (Deep) Reinforcement Learning methods in Python using PyTorch"
 readme = "README.md"
-requires-python = ">=3.12"
+requires-python = ">=3.11"
 dependencies = [
     "numpy>=2.0",
     "torch>=2.0",
